define tasks in tasks directory
